+++ src/store/mutations-types.ts
@@ -12,6 +12,8 @@ export const DELETE_PROJECTS = "DELETE_PROJECTS"
 export const INIT_WORK_DAY = "INIT_WORK_DAY"
 export const FINISH_WORK_DAY = "FINISH_WORK_DAY"
 export const CLEAN_ALL = "CLEAN_ALL"
+export const INCREMENT_TIMER = "INCREMENT_TIMER"
+export const DECREMENT_TIMER = "DECREMENT_TIMER"
 // ALERTS
 export const SHOW_ALERT = "SHOW_ALERT"
 // TRACKERS
